refactor(refinery): extract text classifier helpers to refinery/helpers.rs
